+++ src/test/java/kitchenpos/products/tobe/domain/ProductPriceTest.java
@@ -1,6 +1,6 @@
 package kitchenpos.products.tobe.domain;
 
-import kitchenpos.products.tobe.domain.ProductPrice;
+import kitchenpos.products.tobe.domain.model.ProductPrice;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
